refactor(lpa-questionnaire-web-app): change naming of add supplementary page
