+++ packages/lpa-questionnaire-web-app/src/controllers/supplementary-documents/add-supplementary-document.js
@@ -1,20 +1,20 @@
-const { VIEW } = require('../lib/views');
-const getAppealSideBarDetails = require('../lib/appeal-sidebar-details');
-const { uploadFiles } = require('../lib/file-upload-helpers');
-const { createOrUpdateAppealReply } = require('../lib/appeal-reply-api-wrapper');
-const errorTexts = require('../validators/validation-messages/supplementary-documents-validation-messages');
+const { VIEW } = require('../../lib/views');
+const getAppealSideBarDetails = require('../../lib/appeal-sidebar-details');
+const { uploadFiles } = require('../../lib/file-upload-helpers');
+const { createOrUpdateAppealReply } = require('../../lib/appeal-reply-api-wrapper');
+const errorTexts = require('../../validators/validation-messages/supplementary-documents-validation-messages');
 
 const question = {
-  heading: 'Supplementary planning document',
+  heading: 'Supplementary planning documents',
   section: 'Optional supporting documents',
   title:
-    'Add Supplementary planning document - Appeal Questionnaire - Appeal a householder planning decision - GOV.UK',
+    'Supplementary planning documents - Appeal Questionnaire - Appeal a householder planning decision - GOV.UK',
 };
 
 exports.question = question;
 
 exports.getAddDocument = (req, res) => {
-  // TODO: when list pa[ge is created logic around backlink will need adding. If new page normal backlink is fine
+  // TODO: when list page is created logic around backlink will need adding. If new page normal backlink is fine
   // but if coming from new document button need to set the res.locals.backlink to that page. Session still needed for list page
 
   const backLink = res.locals.backLink || req.session.backLink;
@@ -158,5 +158,5 @@ exports.postAddDocument = async (req, res) => {
     return;
   }
 
-  res.redirect(`/${appealId}/supplementary-documents`);
+  res.redirect(`/${appealId}/supplementary-documents/uploaded-documents`);
 };
